+++ tests/test_touchup_command.rs
@@ -130,7 +130,7 @@ fn test_edit_merge() {
     insta::assert_snapshot!(stdout, @r###"
     Created 9fed260d03e9 merge
     Rebased 1 descendant commits
-    Working copy now at: e80e55af699c (no description set)
+    Working copy now at: b6c2fd3364eb (no description set)
     Added 0 files, modified 0 files, removed 1 files
     "###);
     let stdout = test_env.jj_cmd_success(&repo_path, &["diff", "-s", "-r", "@-"]);